feat: increment mypy coverage from 43 to 58 source files
Add 15 files: code_utils utilities (api_key_storage, path_utils, pytest_utils,
validation, browser_utils, code_utils, numerical_code_utils, diff_utils,
shell_utils, version_check, deduplicate_code, jit_detector), picklepatch,
benchmarking/__init__

Fixes applied:
- diff_utils: fix return type dict[int, int] -> dict[int, float]
- shell_utils: fix return type dict[str, str] -> dict[str, Any]
- version_check: replace _version_cache dict with typed module-level vars
- deduplicate_code: add type annotation for scope_stack
- jit_detector: add type: ignore for ast type narrowing

+++ src/codeflash/code_utils/diff_utils.py
@@ -38,7 +38,7 @@ def normalize_by_max(values: list[float]) -> list[float]:
 
 def create_score_dictionary_from_metrics(
     weights: list[float], *metrics: list[float]
-) -> dict[int, int]:
+) -> dict[int, float]:
     if len(weights) != len(metrics):
         raise ValueError("Number of weights must match number of metrics")
     combined: dict[int, float] = {}
